// web/components/spec/models/Backup.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bckupr
 * Bckupr API
 *
 * The version of the OpenAPI document: latest
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Status } from './Status';
import {
    StatusFromJSON,
    StatusFromJSONTyped,
    StatusToJSON,
} from './Status';
import type { Volume } from './Volume';
import {
    VolumeFromJSON,
    VolumeFromJSONTyped,
    VolumeToJSON,
} from './Volume';

/**
 * 
 * @export
 * @interface Backup
 */
export interface Backup {
    /**
     * 
     * @type {string}
     * @memberof Backup
     */
    id: string;
    /**
     * 
     * @type {Date}
     * @memberof Backup
     */
    created: Date;
    /**
     * 
     * @type {string}
     * @memberof Backup
     */
    type: string;
    /**
     * 
     * @type {Status}
     * @memberof Backup
     */
    status: Status;
    /**
     * 
     * @type {string}
     * @memberof Backup
     */
    error?: string;
    /**
     * 
     * @type {Array<Volume>}
     * @memberof Backup
     */
    volumes: Array<Volume>;
}

/**
 * Check if a given object implements the Backup interface.
 */
export function instanceOfBackup(value: object): boolean {
    if (!('id' in value)) return false;
    if (!('created' in value)) return false;
    if (!('type' in value)) return false;
    if (!('status' in value)) return false;
    if (!('volumes' in value)) return false;
    return true;
}

export function BackupFromJSON(json: any): Backup {
    return BackupFromJSONTyped(json, false);
}

export function BackupFromJSONTyped(json: any, ignoreDiscriminator: boolean): Backup {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'created': (new Date(json['created'])),
        'type': json['type'],
        'status': StatusFromJSON(json['status']),
        'error': json['error'] == null ? undefined : json['error'],
        'volumes': ((json['volumes'] as Array<any>).map(VolumeFromJSON)),
    };
}

export function BackupToJSON(value?: Backup | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'created': ((value['created']).toISOString()),
        'type': value['type'],
        'status': StatusToJSON(value['status']),
        'error': value['error'],
        'volumes': ((value['volumes'] as Array<any>).map(VolumeToJSON)),
    };
}