// web/components/spec/models/Volume.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bckupr
 * Bckupr API
 *
 * The version of the OpenAPI document: latest
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Status } from './Status';
import {
    StatusFromJSON,
    StatusFromJSONTyped,
    StatusToJSON,
} from './Status';

/**
 * 
 * @export
 * @interface Volume
 */
export interface Volume {
    /**
     * 
     * @type {string}
     * @memberof Volume
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Volume
     */
    ext: string;
    /**
     * 
     * @type {string}
     * @memberof Volume
     */
    mount: string;
    /**
     * 
     * @type {Status}
     * @memberof Volume
     */
    status: Status;
    /**
     * 
     * @type {Date}
     * @memberof Volume
     */
    created: Date;
    /**
     * 
     * @type {number}
     * @memberof Volume
     */
    size?: number;
    /**
     * 
     * @type {string}
     * @memberof Volume
     */
    error?: string;
}

/**
 * Check if a given object implements the Volume interface.
 */
export function instanceOfVolume(value: object): boolean {
    if (!('name' in value)) return false;
    if (!('ext' in value)) return false;
    if (!('mount' in value)) return false;
    if (!('status' in value)) return false;
    if (!('created' in value)) return false;
    return true;
}

export function VolumeFromJSON(json: any): Volume {
    return VolumeFromJSONTyped(json, false);
}

export function VolumeFromJSONTyped(json: any, ignoreDiscriminator: boolean): Volume {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'ext': json['ext'],
        'mount': json['mount'],
        'status': StatusFromJSON(json['status']),
        'created': (new Date(json['created'])),
        'size': json['size'] == null ? undefined : json['size'],
        'error': json['error'] == null ? undefined : json['error'],
    };
}

export function VolumeToJSON(value?: Volume | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'name': value['name'],
        'ext': value['ext'],
        'mount': value['mount'],
        'status': StatusToJSON(value['status']),
        'created': ((value['created']).toISOString()),
        'size': value['size'],
        'error': value['error'],
    };
}